refactor: Move all appropriate unit tests to integration test files
